Split up FInventoryMarketHelper
